Fix initialization of menu items
There is a more elegant way to fulfill the items list during the initialization. The current implementation is not only **hm**hm**, but will lead to unpredictable bugs in the future.
in Python: '/' is a float operator. It is a miracle that the code was somehow in working condition.

A simple test, Python 3.11.6:
`type(24/12)`
gives you "float"